Added blob storage check on pre and post cd logs
