+++ src/components/app/details/cdDetails/cd.utils.tsx
@@ -1,6 +1,9 @@
 import React from 'react'
-import { ReactComponent as ArrowDown } from '../../assets/icons/ic-chevron-down.svg'
 import { components } from 'react-select'
+import { ReactComponent as Info } from '../../../../assets/icons/info-filled.svg'
+import { ReactComponent as Question } from '../../../../assets/icons/ic-help-filled.svg'
+import { ReactComponent as OpenInNew } from '../../../../assets/icons/ic-open-in-new.svg'
+import { DOCUMENTATION } from '../../../../config'
 import { multiSelectStyles } from '../../../common'
 
 export const styles = {
@@ -63,4 +66,42 @@ export function Option(props) {
             </div>
         </components.Option>
     )
+}
+
+
+
+const renderLogsNotAvailable = (subtitle?: string): JSX.Element => {
+  return (
+      <div className="flexbox dc__content-center flex-align-center dc__height-inherit">
+          <div>
+          <div className="text-center"><Info className="icon-dim-20"/></div>
+          <div className="text-center cn-0 fs-14 fw-6">Logs not available</div>
+          <div className="text-center cn-0 fs-13 fw-4">{subtitle || 'Logs are available only at runtime.'}</div>
+          </div>
+      </div>
+  )
+}
+
+const renderBlobNotConfigured = (): JSX.Element => {
+  return (
+      <>
+          {renderLogsNotAvailable('Blob storage was not configured at pipeline run.')}
+          <div className="flexbox configure-blob-container pt-8 pr-12 pb-8 pl-12 bcv-1 br-4">
+              <Question className="icon-dim-20" />
+              <span className="fs-13 fw-4 mr-8 ml-8">Want to store logs to view later?</span>
+              <a className="fs-13 fw-6 cb-5 no-decor" href={DOCUMENTATION.ADMIN_PASSWORD} target="_blank">
+                  Configure blob storage
+              </a>
+              <OpenInNew className="icon-dim-20 ml-8" />
+          </div>
+      </>
+  )
+}
+
+export const renderConfigurationError = (isblobStorageConfigured: boolean): JSX.Element => {
+  return (
+      <div className="flexbox dc__content-center flex-align-center dc__height-inherit">
+          {!isblobStorageConfigured ? renderBlobNotConfigured() : renderLogsNotAvailable()}
+      </div>
+  )
 }
